MITMServer backend support integrated; Port Proxy, SSLStrip, and HTTP Cred Harvester functional.
