+++ lib/GUIServer.py
@@ -120,7 +120,7 @@ class dbHandler:
       
       string = base64.b64decode(qstring.split('qstring=')[1].split('&')[0])
       
-      self.conn = sqlite3.connect('/home/rob/Desktop/projects/Subterfuge/attack.db', timeout=1) #You like the dick, you stupid variable python path piece of crap... !
+      self.conn = sqlite3.connect('/home/adhd/Desktop/projects/Subterfuge/attack.db', timeout=1) #You like the dick, you stupid variable python path piece of crap... !
       self.conn.execute('pragma foreign_keys = on')
       self.conn.commit()
       self.cur = self.conn.cursor()
