+++ lib/metafeeds.py
@@ -1,5 +1,5 @@
 class sslstrip:
-   def __init__( self):
+   def __init__(self):
       import os
       import sys
       sys.path.append("../")
@@ -9,10 +9,16 @@ class sslstrip:
       import sys
       sys.path.append("../")
       import subprocess
+      
+      print "Starting SSLStrip on " + PROXY_PORT
+      
       package_dir = os.path.dirname(os.path.dirname(__file__)) + '/packages/'
       print package_dir
+      
+      
+      
       process = subprocess.Popen('python ' + package_dir + 'sslstrip.py -w ' + package_dir + '/sslstrip.log -l ' + str(PROXY_PORT) + ' -f &', shell=True)
       
-      print process.pid
+      return process.pid
       
       #os.system('python packages/sslstrip.py -w packages/sslstrip.log -l ' + str(PROXY_PORT) + ' -f &')
