+++ packages/harvester/passwordfields.lst
@@ -1,0 +1,90 @@
+acct_pass
+apass
+auth_passwd
+Apassword
+A_Password
+cppw
+credential_1
+ctl00%24ContentInfo%24SignIn%24password%24txtPassword
+cugPassword
+cuserpass
+data_pass
+emagicpasswd
+form_password
+form_passwort
+form_pw
+fpass
+frmUserPass
+gallerypassword
+geckos
+hp_pwd
+horde_pass
+inspass
+inputPassword
+intCodSegreto
+in_passwort
+iPassword
+j_password
+kennwort
+login2
+login_pass
+login_password
+login_passwort
+login-password
+logonpassword
+logonPwd
+logon_pwd
+log_pass
+l_password
+LoginForm_Password
+Loginpass
+machpasswort
+member_PASS
+my_pass
+newpassword
+newLoginPassword
+nutzer_password
+pass
+passform
+passwd
+password
+passwort
+passwortm
+passwrd
+paswwort
+pin
+psw
+pw
+pwd
+pwlogin
+p_key
+p_passwd
+Passwd
+Password
+Password_Textbox
+secretkey
+securityPassword
+segreto
+session%5Bpassword%5D
+session_password
+theAccountPW
+txtPassword
+txtPasswort
+upassword
+upasswort
+user%5Bpassword%5D
+userpass
+userpassword
+userpwd
+user_password
+user_passwort
+user_security_password
+use_pwd
+UPW
+U_Pass
+vb_login_password
+webpass
+wpPassword
+YaBBpassword
+zpassword
+
