+++ packages/harvester/usernamefields.lst
@@ -1,0 +1,113 @@
+aadvantageNumber
+accountname
+acctname
+acct_name
+alias
+anwender
+auser
+auth_userId
+Anick
+A_userName
+benutzer
+Benutzername
+bor_id
+cardNumber
+codice
+cpid
+credential
+credential_0
+ctl00%24ContentInfo%24SignIn%24onepass%24txtField
+cugUsername
+cuser
+data_nickname
+emagicloginid
+email
+emailn
+email_address
+Email
+EmailAddress
+Email_Texbox
+fmembernr
+form_loginname
+form_userid
+form_userlogin
+frmUserID
+f_user
+gecko
+horde_user
+hp_user_name
+id
+ident
+inputEmailHandle
+insname
+in_benutzername
+iLogin
+I_username
+j_username
+kdid
+kdnr
+kennung
+log
+log_name
+log_nummer
+login
+login1
+loginid
+loginId
+loginname
+login_email
+login_id
+login_name
+login_username
+login-appleId
+logonEmailId
+logon_name
+logon_user
+LoginForm_Login
+machlogin
+mailaddress
+matchcode
+member
+membername
+member_ID
+my_user
+nick
+nickform
+nickname
+newusername
+newLoginName
+NAME
+pop_login
+o_uid
+p_user
+screenname
+session%5Busername_or_email%5D
+session_key
+sign-in
+spielername
+strCodUtente
+theAccountName
+txtlogin
+txtUsername
+txtUserID
+ulogin
+uname
+use_username
+user
+userid
+user%5Bemail%5D
+user_email
+user_id
+user_login
+user_mail
+user_name
+username
+userName
+UserName
+U_Name
+vb_login_username
+webuser
+wpName
+YaBBusername
+z_name
+
